backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
pydantic[email]
python-dotenv
httpx
pytest
pytest-asyncio
redis
PyJWT>=2.8.0
python-multipart